// lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      applications: {
        Row: {
          created_at: string
          id: string
          institution_id: string
          status: string
          student_id: string
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          id?: string
          institution_id: string
          status?: string
          student_id: string
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          id?: string
          institution_id?: string
          status?: string
          student_id?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "applications_institution_id_fkey"
            columns: ["institution_id"]
            isOneToOne: false
            referencedRelation: "institutions"
            referencedColumns: ["id"]
          },
        ]
      }
      institutions: {
        Row: {
          address: string | null
          board: string | null
          city: string | null
          classes_offered: string[] | null
          contact_email: string | null
          contact_phone: string | null
          created_at: string
          exams_coached: string[] | null
          features: Json | null
          fee_structure: string | null
          id: string
          images: string[] | null
          is_approved: boolean
          name: string | null
          pincode: string | null
          programs_offered: string[] | null
          state: string | null
          student_teacher_ratio: number | null
          type: string | null
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          address?: string | null
          board?: string | null
          city?: string | null
          classes_offered?: string[] | null
          contact_email?: string | null
          contact_phone?: string | null
          created_at?: string
          exams_coached?: string[] | null
          features?: Json | null
          fee_structure?: string | null
          id?: string
          images?: string[] | null
          is_approved?: boolean
          name?: string | null
          pincode?: string | null
          programs_offered?: string[] | null
          state?: string | null
          student_teacher_ratio?: number | null
          type?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          address?: string | null
          board?: string | null
          city?: string | null
          classes_offered?: string[] | null
          contact_email?: string | null
          contact_phone?: string | null
          created_at?: string
          exams_coached?: string[] | null
          features?: Json | null
          fee_structure?: string | null
          id?: string
          images?: string[] | null
          is_approved?: boolean
          name?: string | null
          pincode?: string | null
          programs_offered?: string[] | null
          state?: string | null
          student_teacher_ratio?: number | null
          type?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never